fix(sms): sending and saving to multiple recipients
A dict does not work for the phone numbers here, because all the
"keys" are named "Phone". Therefore we ended up using only the last
number in the given list.

Pass a single-key dict and a list for its values so xmltodict.unparse
serializes it as multiple "Phone" elements which is what we want.

Regression in 1.5, 0b99acd0226de29bc6d549858a27711fc538611d